+++ provision/stack_creator.go
@@ -43,14 +43,15 @@ type (
 
 		stackName string
 
-		serviceVersion string
-
 		config            conf.Config
 		environment       conf.Environment
 		region            conf.Region
 		servicePayloadKey string
+		secretPayloadKey  string
 
 		secretsKey string
+
+		registryDeployment bool
 
 		roleSession *session.Session
 
@@ -69,14 +70,14 @@ const (
 
 func (recv *stackCreator) createUpdateStackForRegion(outChan chan CreateStackRegionOutput, errChan chan struct{}) {
 
-	_, success := recv.uploadServicePayload()
+	checksum, success := recv.uploadServicePayload()
 	if !success {
 		// uploadServicePayload logs errors. all we care about is success
 		errChan <- struct{}{}
 		return
 	}
 
-	if !recv.uploadSecrets() {
+	if !recv.uploadSecrets(checksum) {
 		// uploadSecrets logs errors. all we care about is success
 		errChan <- struct{}{}
 		return
@@ -154,6 +155,59 @@ func (recv *stackCreator) uploadServicePayload() (checksum string, success bool)
 		return
 	}
 
+	dockerRegistry := os.Getenv(constants.EnvDockerRegistry)
+	dockerRepository := os.Getenv(constants.EnvDockerRepository)
+	dockerPushUsername := os.Getenv(constants.EnvDockerPushUsername)
+	dockerPushPassword := os.Getenv(constants.EnvDockerPushPassword)
+
+	if dockerRegistry != "" && dockerRepository != "" {
+		recv.registryDeployment = true
+
+		if dockerPushUsername != "" && dockerPushPassword != "" {
+
+			recv.log.Info("docker login")
+			loginCmd := exec.Command("docker", "login",
+				"-u", dockerPushUsername,
+				"-p", dockerPushPassword,
+				dockerRegistry)
+			loginCmd.Stderr = os.Stderr
+			err := loginCmd.Run()
+			if err != nil {
+				recv.log.Error("docker login", "Error", err)
+				return
+			}
+		}
+
+		containerCount := len(recv.region.Containers)
+		pushSuccessChan := make(chan bool, containerCount)
+
+		for _, container := range recv.region.Containers {
+
+			go func(log log15.Logger, container conf.Container) {
+				log = log.New("ImageTag", container.Name)
+
+				log.Info("docker push")
+				pushCmd := exec.Command("docker", "push", container.Name)
+				pushCmd.Stderr = os.Stderr
+				err := pushCmd.Run()
+				if err != nil {
+					log.Error("docker push", "Error", err)
+					pushSuccessChan <- false
+					return
+				}
+				pushSuccessChan <- true
+
+			}(recv.log, *container)
+		}
+
+		for i := 0; i < containerCount; i++ {
+			pushSuccess := <-pushSuccessChan
+			if !pushSuccess {
+				return
+			}
+		}
+	}
+
 	success = true
 	return
 }
@@ -182,6 +236,11 @@ func (recv *stackCreator) createStack() (stackId string, success bool) {
 		Bucket: aws.String(recv.region.S3Bucket),
 		Key:    aws.String(templateS3Key),
 		Body:   bytes.NewReader(templateBytes),
+	}
+
+	if recv.region.SSEKMSKeyId != nil {
+		uploadInput.SSEKMSKeyId = recv.region.SSEKMSKeyId
+		uploadInput.ServerSideEncryption = aws.String("aws:kms")
 	}
 
 	s3Manager := s3manager.NewUploader(recv.roleSession)
@@ -289,5 +348,5 @@ func (recv *stackCreator) s3KeyRoot(prefixOpt int) string {
 	}
 
 	return fmt.Sprintf("%s/%s/%s/%s",
-		prefix, recv.config.ServiceName, recv.environment.Name, recv.serviceVersion)
+		prefix, recv.config.ServiceName, recv.environment.Name, recv.config.ServiceVersion)
 }